labs/04-deploy-ai/02-chainlitagent-ui/requirements.txt
chainlit==0.6.402
fastapi==0.99.1
openai==0.27.9
langchain==0.0.271
python-dotenv==1.0.0